+++ TODO.md
@@ -4,5 +4,5 @@
 ## Additional Features:
 - File uploads by drag/dropping into terminal
 - Rendering of hyperlinks/citations, at least for claude models
-- Ability to switch models mid-REPL session with a command, keeping history
+- Ability to switch models mid-session with a command, keeping history
 - Add precommit hooks
