+++ cmd/run.go
@@ -5,7 +5,7 @@ package cmd
 
 import (
 	"github.com/gregriff/gpt-cli-go/models/anthropic"
-	"github.com/gregriff/gpt-cli-go/repl"
+	"github.com/gregriff/gpt-cli-go/tui"
 	"github.com/spf13/cobra"
 )
 
@@ -20,7 +20,7 @@ var runCmd = &cobra.Command{
 	PreRunE: func(cmd *cobra.Command, args []string) error {
 		return anthropic.ValidateModelName(modelName)
 	},
-	Run: runRepl,
+	Run: runTUI,
 }
 
 func init() {
@@ -38,8 +38,8 @@ func init() {
 	// runCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
 }
 
-func runRepl(cmd *cobra.Command, args []string) {
+func runTUI(cmd *cobra.Command, args []string) {
 	systemPrompt := "You are a concise assistant to a software engineer"
-	repl := repl.NewREPL(systemPrompt, modelName, 2048)
-	repl.Start()
+	tui := tui.NewTUI(systemPrompt, modelName, 2048)
+	tui.Start()
 }
