+++ tui/text.go
@@ -1,4 +1,4 @@
-package repl
+package tui
 
 import "unicode"
